Isolate agency::bulk_async within its own header
