+++ agency/detail/bulk_functions/bulk_async_execution_policy.hpp
@@ -5,6 +5,8 @@
 #include <agency/detail/type_traits.hpp>
 #include <agency/detail/bulk_functions/execute_agent_functor.hpp>
 #include <agency/detail/bulk_functions/single_result.hpp>
+#include <agency/detail/bulk_functions/bulk_invoke_execution_policy.hpp>
+#include <agency/detail/execution_policy_traits.hpp>
 #include <agency/execution_agent.hpp>
 #include <agency/executor_traits.hpp>
 #include <utility>
@@ -28,33 +30,13 @@ template<class ExecutionPolicy, class Function, class... Args>
 using bulk_async_execution_policy_result_t = typename bulk_async_execution_policy_result<ExecutionPolicy,Function,Args...>::type;
 
 
-template<bool enable, class ExecutionPolicy, class Function, class... Args>
-struct enable_if_bulk_async_execution_policy_impl {};
-
-template<class ExecutionPolicy, class Function, class... Args>
-struct enable_if_bulk_async_execution_policy_impl<true, ExecutionPolicy, Function, Args...>
-{
-  using type = bulk_async_execution_policy_result_t<ExecutionPolicy,Function,Args...>;
-};
-
-template<class ExecutionPolicy, class Function, class... Args>
-struct enable_if_bulk_async_execution_policy
-  : enable_if_bulk_async_execution_policy_impl<
-      is_bulk_call_possible_via_execution_policy<decay_t<ExecutionPolicy>,Function,Args...>::value,
-      decay_t<ExecutionPolicy>,
-      Function,
-      Args...
-    >
-{};
-
-
 template<size_t... UserArgIndices, size_t... SharedArgIndices, class ExecutionPolicy, class Function, class... Args>
 bulk_async_execution_policy_result_t<
   ExecutionPolicy, Function, Args...
 >
-  bulk_async_execution_policy_impl(index_sequence<UserArgIndices...>,
-                                   index_sequence<SharedArgIndices...>,
-                                   ExecutionPolicy& policy, Function f, Args&&... args)
+  bulk_async_execution_policy(index_sequence<UserArgIndices...>,
+                              index_sequence<SharedArgIndices...>,
+                              ExecutionPolicy& policy, Function f, Args&&... args)
 {
   using agent_type = typename ExecutionPolicy::execution_agent_type;
   using agent_traits = execution_agent_traits<agent_type>;
@@ -81,19 +63,6 @@ bulk_async_execution_policy_result_t<
 
 
 } // end detail
-
-
-template<class ExecutionPolicy, class Function, class... Args>
-typename detail::enable_if_bulk_async_execution_policy<
-  ExecutionPolicy, Function, Args...
->::type
-  bulk_async(ExecutionPolicy&& policy, Function f, Args&&... args)
-{
-  using agent_traits = execution_agent_traits<typename std::decay<ExecutionPolicy>::type::execution_agent_type>;
-  const size_t num_shared_params = detail::execution_depth<typename agent_traits::execution_category>::value;
-
-  return detail::bulk_async_execution_policy_impl(detail::index_sequence_for<Args...>(), detail::make_index_sequence<num_shared_params>(), policy, f, std::forward<Args>(args)...);
-}
 
 
 } // end agency
