+++ testing/bulk_async/cuda/device_lambda/scope_result.cu
@@ -1,5 +1,4 @@
-#include <agency/execution_policy.hpp>
-#include <agency/bulk_invoke.hpp>
+#include <agency/bulk_async.hpp>
 #include <agency/cuda/execution_policy.hpp>
 #include <cassert>
 #include <vector>
