+++ testing/bulk_async/void_result.cpp
@@ -1,3 +1,4 @@
+#include <agency/bulk_async.hpp>
 #include <agency/execution_policy.hpp>
 #include <atomic>
