+++ testing/unorganized/main.cpp
@@ -1,5 +1,6 @@
-#include <iostream>
+#include <agency/bulk_async.hpp>
 #include <agency/execution_policy.hpp>
+#include <iostream>
 
 std::mutex mut;
