+++ testing/unorganized/test.cpp
@@ -1,5 +1,6 @@
 #include <iostream>
 #include <mutex>
+#include <agency/bulk_async.hpp>
 #include <agency/execution_policy.hpp>
 
 int main()
