+++ testing/unorganized/test_concurrent.cpp
@@ -1,4 +1,5 @@
 #include <iostream>
+#include <agency/bulk_async.hpp>
 #include <agency/execution_policy.hpp>
 
 int main()
